+++ MapboxSearch/common-tests/src/main/java/com/mapbox/search/common/tests/CoreTestDataFactory.kt
@@ -48,6 +48,7 @@ fun createTestCoreSearchOptions(
     timeDeviation: Double? = null,
     addonAPI: Map<String, String>? = null,
     offlineSearchPlacesOutsideBbox: Boolean = false,
+    ensureResultsPerCategory: Boolean? = null,
 ): SearchOptions = SearchOptions(
     proximity,
     origin,
@@ -67,6 +68,7 @@ fun createTestCoreSearchOptions(
     timeDeviation,
     addonAPI?.let { it as? HashMap<String, String> ?: HashMap(it) },
     offlineSearchPlacesOutsideBbox,
+    ensureResultsPerCategory,
 )
 
 @Suppress("LongParameterList")