+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog for the Mapbox Search SDK for Android
 
+## 2.7.0-rc.1
+
+### New features
+- Added new function to search for multiple categories, e.g., `SearchEngine.search(listOf("coffee_shop", "hotel"), ...)`,
+- New option `CategorySearchOptions.ensureResultsPerCategory` is available. It allows to request category search results to include at least one POI for each category, provided a POI is available in a nearby location.
+
+### Bug fixes
+- Fixed photos metadata parsing for the `ApiType.SEARCH_BOX`.
+
+### Mapbox dependencies
+- Search Native SDK `2.7.0-rc.1`
+- Common SDK `24.9.0-rc.1`
+
+
+
 ## 2.7.0-beta.1
 
 ### New features